feat: put XmlToCsvUploader for XML and CsvUploader in separate files
